+++ api/internal/v2/ocr/types/hint.go
@@ -14,7 +14,7 @@ type HintRequest struct {
 	TaskStruct    TaskStruct `json:"task_struct"`
 	RawTaskText   string     `json:"raw_task_text"`
 	Level         HintLevel  `json:"level"`                    // "L1" | "L2" | "L3"
-	Grade         *int       `json:"grade"`                    // 1..4
+	Grade         *int64     `json:"grade"`                    // 1..4
 	PreviousHints []string   `json:"previous_hints,omitempty"` // max 2 (валидируется схемой)
 	Locale        string     `json:"locale,omitempty"`         // "ru-RU" | "en-US"
 }
